Move pregnancy services to pregnancyInfoService from evaluateFhirDataService. Move tests
