+++ containers/ecr-viewer/src/app/view-data/components/EcrDocument/accordion-items.tsx
@@ -19,9 +19,9 @@ import {
   evaluateProviderData,
   evaluateFacilityData,
   evaluateHospitalEncounterData,
-  evaluatePregnancyData,
 } from "@/app/view-data/services/evaluateFhirDataService";
 import { evaluateDemographicsData } from "@/app/view-data/services/demographicsService";
+import { evaluatePregnancyData } from "@/app/view-data/services/pregnancyInfoService";
 import { evaluateLabInfoData } from "@/app/view-data/services/labsService";
 import {
   FhirIndex,
